Storage/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
bigtree==0.22.3
bson==0.5.10
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
confluent-kafka==2.6.0
contourpy==1.3.1
cycler==0.12.1
dask==2024.11.2
dask-expr==1.1.19
debugpy==1.8.8
decorator==5.1.1
dill==0.3.9
executing==2.1.0
fastapi==0.115.5
feast==0.41.3
flatbuffers==24.3.25
fonttools==4.55.2
fsspec==2024.10.0
gast==0.6.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
hiredis==2.4.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.14.0
kiwisolver==1.4.7
libclang==18.1.1
locket==1.0.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
ml-dtypes==0.2.0
mmh3==5.0.1
mypy==1.13.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==4.6.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
rsa==4.9
scikit-learn==1.4.0
scipy==1.11.3
seaborn==0.12.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.2
statsmodels==0.14.0
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.14.0
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typeguard==4.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvicorn-worker==0.2.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.21.0
